Se agregan algunas validaciones a la hora de encontrar un articulo por ID o un lote registrado.Tambien se logra actualizar las unidades disponibles de un lote

+++ Difficult-SpringBoot App/difficultApp/src/main/kotlin/ar/com/phm2022/aplicacion/controller/UsuarioController.kt
@@ -19,7 +19,7 @@ class ItemJson{
     var idItemJson:Long=0
     var idArticulo:Long = 0
     var cantidad:Int = 0
-    var loteElegido:Int = 0
+    var loteElegido:Long = 0
 }
 
 @RestController
